% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_rf.R
\name{snp_rf}
\alias{snp_rf}
\title{Random Forests for Final SNP Selection}
\usage{
snp_rf(data_list, y)
}
\arguments{
\item{data_list}{Object obtained from the \code{elasticnet} function in this
package.}

\item{y}{A numeric vector of phenotypes for the data. Can be either
continuous or binary.}
}
\value{
\code{snp_rf} returns a list for each dataset in
\code{data_list}. Each element of the list contains
the variable importance measures, variable importance plot,
and the alpha value associated with the \code{elasticnet} model used
to create the dataset. Each item is called using the list number
and name, such as results[[1]]$vi

\item{vi}{Vector that includes the SNPs and their importance.}
\item{vi_plot}{Variable importance plot.}
\item{alpha}{Alpha value from the elastic net model used to trim the
input data.}
}
\description{
Computes the importance of each SNP in the input datasets and creates
variable importance plots for final SNP selection using
random forests. The function
uses the dataset list returned by the \code{elasticnet} function
as input.
}
