% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_tree.R
\name{snp_tree}
\alias{snp_tree}
\title{CART for Final SNP Selection}
\usage{
snp_tree(data_list, y)
}
\arguments{
\item{data_list}{Object obtained from the \code{elasticnet} function.}

\item{y}{A numeric vector of phenotypes for the data.}
}
\value{
\code{snp_tree} returns a list for each dataset in \code{data_list}.
Each of the sublists
contains the pruned tree created by rpart, a vector containing the SNPs where
the tree splits, and the alpha value
associated with the elastic net model used to create the dataset. Each
item is called using the list number
and name, such as results[[1]]$tree and results[[1]]$snps.

\item{tree}{Tree produced by rpart. Tree is pruned using the 1-se rule.}
\item{snps}{Vector of SNPs where the tree splits. }
\item{alpha}{Alpha value from the elastic net model used to trim the
input data.}
}
\description{
Creates classification and regression trees for final SNP selection using
rpart. The function uses the dataset list from the \code{elasticnet}
function in this package as input and returns a list of trees and the
SNPs where the trees split. The trees are pruned using the 1-se rule.
}
