% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_plot.R
\name{tree_plot}
\alias{tree_plot}
\title{Plot Trees from Final SNP Selection}
\usage{
tree_plot(tree, nrow = 2, ncol = 4, palette = "BlGnYl")
}
\arguments{
\item{tree}{Object obtained from the \code{snp_tree} function.}

\item{nrow, ncol}{Integers that specify the number of rows and
columns in the plotting screen.}

\item{palette}{User specified palette using options from \code{rpart.plot}.}
}
\value{
\code{plot_tree} plots the trees obtained from \code{snp_tree} using
the \code{rpart.plot} package. If a plot is missing for a value of
alpha, it is because one or no variables were retained from the elastic net
model.
}
\description{
Plots the trees obtained using \code{snp_tree}.
}
